(PC-33462) refactor(venue): extract type in types file and remove as
