+++ src/features/offer/components/MoviesScreeningCalendar/MoviesScreeningCalendar.tsx
@@ -7,7 +7,7 @@ import { UseRouteType } from 'features/navigation/RootNavigator/types'
 import { MovieCalendarProvider } from 'features/offer/components/MoviesScreeningCalendar/MovieCalendarContext'
 import { VenueCalendar } from 'features/offer/components/MoviesScreeningCalendar/VenueCalendar'
 import { OfferTileWrapper } from 'features/offer/components/OfferTile/OfferTileWrapper'
-import { VenueOffers } from 'features/venue/api/useVenueOffers'
+import { VenueOffers } from 'features/venue/types'
 import { useFeatureFlag } from 'libs/firebase/firestore/featureFlags/useFeatureFlag'
 import { RemoteStoreFeatureFlags } from 'libs/firebase/firestore/types'
 import { getDates } from 'shared/date/getDates'
