+++ src/features/venue/components/VenueBody/VenueBody.tsx
@@ -4,13 +4,13 @@ import { useTheme } from 'styled-components/native'
 
 import { VenueResponse } from 'api/gen'
 import { GtlPlaylistData } from 'features/gtlPlaylist/types'
-import type { VenueOffers as VenueOffersType } from 'features/venue/api/useVenueOffers'
 import { PracticalInformation } from 'features/venue/components/PracticalInformation/PracticalInformation'
 import { TabLayout } from 'features/venue/components/TabLayout/TabLayout'
 import { VENUE_CTA_HEIGHT_IN_SPACES } from 'features/venue/components/VenueCTA/VenueCTA'
 import { VenueMessagingApps } from 'features/venue/components/VenueMessagingApps/VenueMessagingApps'
 import { VenueOffers } from 'features/venue/components/VenueOffers/VenueOffers'
 import { VenueThematicSection } from 'features/venue/components/VenueThematicSection/VenueThematicSection'
+import type { VenueOffers as VenueOffersType } from 'features/venue/types'
 import { Tab } from 'features/venue/types'
 import { analytics } from 'libs/analytics'
 import { SectionWithDivider } from 'ui/components/SectionWithDivider'
