+++ src/features/venue/components/VenueContent/VenueContent.tsx
@@ -8,7 +8,6 @@ import { VenueResponse, VenueTypeCodeKey } from 'api/gen'
 import { GtlPlaylistData } from 'features/gtlPlaylist/types'
 import { CineContentCTA } from 'features/offer/components/OfferCine/CineContentCTA'
 import { useOfferCTA } from 'features/offer/components/OfferContent/OfferCTAProvider'
-import { VenueOffers } from 'features/venue/api/useVenueOffers'
 import { VenueBody } from 'features/venue/components/VenueBody/VenueBody'
 import { VenueCTA } from 'features/venue/components/VenueCTA/VenueCTA'
 import { VenueHeader } from 'features/venue/components/VenueHeader/VenueHeader'
@@ -16,6 +15,7 @@ import { VenueTopComponent } from 'features/venue/components/VenueTopComponent/V
 import { VenueWebMetaHeader } from 'features/venue/components/VenueWebMetaHeader'
 import { VideoSection } from 'features/venue/components/VideoSection/VideoSection'
 import { VENUE_VIDEO_FAKEDOOR_DATA } from 'features/venue/constants'
+import { VenueOffers } from 'features/venue/types'
 import { analytics, isCloseToBottom } from 'libs/analytics'
 import { useRemoteConfigContext } from 'libs/firebase/remoteConfig/RemoteConfigProvider'
 import { useFunctionOnce } from 'libs/hooks'
